Moved events from attribute panel to configure panel
Moved events from attribute panel to configure panel
